ai-network-autoconfig/requirements.txt
netmiko
pyyaml
# openai  <-- Commented out because we are using the Dummy AI